Provided actual file that was used Summer 2015 for processing and inserting the records into the database. Need to add server validation.
Added error_log.txt to show errors on our PeopleSoft stored procedure when
there was more than 1 result. Didn't happen all the time, but this error
caused an error when pulling particular students and checking them with
those that had a valid check-in date and a completed housing application
on file.